[#70] refactor: implement base64 as function
